+++ CHANGES.rst
@@ -1,9 +1,10 @@
 Version 0.2.0
 -------------
 
-Released 2025-01-04
+Released 2025-01-08
 
 -   Add dyndns service.
+-   Add unbound service.
 
 Version 0.1.0
 -------------